organize interfaces into util
