+++ src/shared/utils/app/get-data-type-string.ts
@@ -1,4 +1,4 @@
-import { DataType } from "../../interfaces";
+import { DataType } from "@utils/types";
 
 export default function getDataTypeString(dt: DataType) {
   return dt === DataType.Post ? "Post" : "Comment";
