+++ src/shared/utils/types/ban-type.ts
@@ -1,0 +1,4 @@
+export enum BanType {
+  Community,
+  Site,
+}
