+++ src/shared/utils/types/comment-node-i.ts
@@ -1,0 +1,7 @@
+import { CommentView } from "lemmy-js-client";
+
+export default interface CommentNodeI {
+  comment_view: CommentView;
+  children: Array<CommentNodeI>;
+  depth: number;
+}
