+++ src/shared/utils/types/comment-view-type.ts
@@ -1,0 +1,4 @@
+export enum CommentViewType {
+  Tree,
+  Flat,
+}
