+++ src/shared/utils/types/data-type.ts
@@ -1,0 +1,4 @@
+export enum DataType {
+  Post,
+  Comment,
+}
