+++ src/shared/utils/types/i-lemmy-config.ts
@@ -1,0 +1,3 @@
+export default interface ILemmyConfig {
+  wsHost?: string;
+}
