+++ src/shared/utils/types/iso-data-optional-site.ts
@@ -1,0 +1,6 @@
+import { IsoData, RouteData } from "@utils/types";
+
+export type IsoDataOptionalSite<T extends RouteData = any> = Partial<
+  IsoData<T>
+> &
+  Pick<IsoData<T>, Exclude<keyof IsoData<T>, "site_res">>;
