+++ src/shared/utils/types/person-details-view.ts
@@ -1,0 +1,6 @@
+export enum PersonDetailsView {
+  Overview = "Overview",
+  Comments = "Comments",
+  Posts = "Posts",
+  Saved = "Saved",
+}
