+++ src/shared/utils/types/post-form-params.ts
@@ -1,0 +1,5 @@
+export default interface PostFormParams {
+  name?: string;
+  url?: string;
+  body?: string;
+}
