+++ src/shared/utils/types/purge-type.ts
@@ -1,0 +1,6 @@
+export enum PurgeType {
+  Person,
+  Community,
+  Post,
+  Comment,
+}
