+++ src/shared/utils/types/route-data.ts
@@ -1,0 +1,3 @@
+import { RequestState } from "../../services/HttpService";
+
+export type RouteData = Record<string, RequestState<any>>;
