+++ src/shared/utils/types/vote-type.ts
@@ -1,0 +1,4 @@
+export enum VoteType {
+  Upvote,
+  Downvote,
+}
